deployment/aurora-standard-node/services/training-scheduler/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
httpx==0.25.2
celery==5.3.4
sqlalchemy==2.0.23
